feat(recording): add fixed-lag saved track smoothing
